// admin-service/src/dto/vendor.dto.ts
export class VendorDto{
    name: string;
    email: string;
    OTP: string;
    password: string;  
    oldpassword : string;
    newpassword : string;
    vendor_id: string;
    countryCode:string;
    createdAt: Date;
    updatedAt: Date;
    profileImg: string;
    contact: string;
    status:string;
    pageNo: string;
    perPage: string;
    search: string;
    toDate: string;
    fromDate: string;
    language:string;
    type:string
  
}